requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.3.0
click==8.1.2
coverage==6.3.2
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
flake8==4.0.1
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
